Refactor: modify homepage design (#303)
