+++ src/app/(route)/home/HomePage.tsx
@@ -1,12 +1,12 @@
 'use client';
 
 import GreetingBlock from './_components/GreetingBlock/GreetingBlock';
-import HomeQueryBox from './_components/HomeQueryBox/HomeQueryBox';
+import HomeQueryBox from './_components/HomeQueryBox';
 
 export default function Home() {
   return (
     <div className="flex h-screen w-full flex-col items-center justify-center">
-      <div className="mb-7 flex w-fit flex-col">
+      <div className="mb-7 ml-4 flex w-full max-w-184 flex-col">
         <div className="font-title-md ml-6 flex flex-col gap-1.5 md:flex-row">
           <span>저장한 링크 속 내용을 바탕으로</span>
           <span> 답변해 드려요.</span>
