+++ src/app/(route)/home/_components/HomeQueryBox.tsx
